+++ workforce/schedule/public-read-nonce-guard.js
@@ -1,37 +1,25 @@
-const NONCE_STATE_KEY = 'schedule-public-read-nonce'
+import { DurableObject } from 'cloudflare:workers'
 
-function validExpiry(value, now) {
-  return Number.isSafeInteger(value)
-    && value > now
-    && value <= now + (15 * 60 * 1000)
-}
+import {
+  clearSchedulePublicReadNonce,
+  consumeSchedulePublicReadNonce,
+} from './public-read-nonce-state.js'
 
 /**
  * One Durable Object is deterministically assigned to every authenticated
  * nonce. Durable Object event serialization makes consuming that nonce an
  * atomic operation, so a duplicated signed request cannot pass concurrently.
  */
-export class SchedulePublicReadNonceGuard {
-  constructor(state) {
-    this.state = state
+export class SchedulePublicReadNonceGuard extends DurableObject {
+  constructor(ctx, env) {
+    super(ctx, env)
   }
 
-  async consume({ expiresAt, now = Date.now() } = {}) {
-    if (!validExpiry(expiresAt, now)) return { ok: false, code: 'INVALID_EXPIRY' }
-
-    const previousExpiry = await this.state.storage.get(NONCE_STATE_KEY)
-    if (Number.isSafeInteger(previousExpiry) && previousExpiry > now) {
-      return { ok: false, code: 'REPLAYED' }
-    }
-
-    await this.state.storage.put(NONCE_STATE_KEY, expiresAt)
-    await this.state.storage.setAlarm(expiresAt)
-    return { ok: true }
+  async consume({ expiresAt } = {}) {
+    return consumeSchedulePublicReadNonce(this.ctx.storage, { expiresAt })
   }
 
   async alarm() {
-    await this.state.storage.delete(NONCE_STATE_KEY)
+    await clearSchedulePublicReadNonce(this.ctx.storage)
   }
 }
-
-export const __testables = { NONCE_STATE_KEY, validExpiry }
