+++ workforce/schedule/public-read-nonce-guard.test.js
@@ -1,7 +1,11 @@
 import assert from 'node:assert/strict'
 import test from 'node:test'
+import { readFileSync } from 'node:fs'
 
-import { SchedulePublicReadNonceGuard } from './public-read-nonce-guard.js'
+import {
+  clearSchedulePublicReadNonce,
+  consumeSchedulePublicReadNonce,
+} from './public-read-nonce-state.js'
 
 function createState() {
   const entries = new Map()
@@ -19,25 +23,31 @@ function createState() {
 
 test('nonce guard consumes an authenticated nonce once and removes it after its alarm', async () => {
   const state = createState()
-  const guard = new SchedulePublicReadNonceGuard(state)
   const now = Date.now()
   const expiresAt = now + 60_000
 
-  assert.deepEqual(await guard.consume({ expiresAt, now }), { ok: true })
-  assert.deepEqual(await guard.consume({ expiresAt, now: now + 1 }), { ok: false, code: 'REPLAYED' })
+  assert.deepEqual(await consumeSchedulePublicReadNonce(state.storage, { expiresAt, now }), { ok: true })
+  assert.deepEqual(await consumeSchedulePublicReadNonce(state.storage, { expiresAt, now: now + 1 }), { ok: false, code: 'REPLAYED' })
   assert.equal(state.snapshot().alarmAt, expiresAt)
 
-  await guard.alarm()
-  assert.deepEqual(await guard.consume({ expiresAt: now + 120_000, now: now + 60_001 }), { ok: true })
+  await clearSchedulePublicReadNonce(state.storage)
+  assert.deepEqual(await consumeSchedulePublicReadNonce(state.storage, { expiresAt: now + 120_000, now: now + 60_001 }), { ok: true })
 })
 
 test('nonce guard refuses malformed or unbounded expiries without retaining state', async () => {
   const state = createState()
-  const guard = new SchedulePublicReadNonceGuard(state)
   const now = Date.now()
 
   for (const expiresAt of [undefined, now, now - 1, now + (16 * 60 * 1000), 'not-a-number']) {
-    assert.deepEqual(await guard.consume({ expiresAt, now }), { ok: false, code: 'INVALID_EXPIRY' })
+    assert.deepEqual(await consumeSchedulePublicReadNonce(state.storage, { expiresAt, now }), { ok: false, code: 'INVALID_EXPIRY' })
   }
   assert.equal(state.snapshot().entries.size, 0)
+})
+
+test('nonce guard is a Workers Durable Object with an RPC-compatible base constructor', () => {
+  const source = readFileSync(new URL('./public-read-nonce-guard.js', import.meta.url), 'utf8')
+  assert.match(source, /import\s+\{\s*DurableObject\s*\}\s+from\s+['"]cloudflare:workers['"]/)
+  assert.match(source, /export\s+class\s+SchedulePublicReadNonceGuard\s+extends\s+DurableObject/)
+  assert.match(source, /constructor\(ctx, env\)\s*\{\s*super\(ctx, env\)/)
+  assert.match(source, /async\s+consume\(/)
 })
